Reorganize settings-related directory subtree
