+++ lib/skate_web/controllers/page_controller.ex
@@ -1,7 +1,7 @@
 defmodule SkateWeb.PageController do
   require Logger
   use SkateWeb, :controller
-  alias Skate.Settings
+  alias Skate.Settings.UserSettings
   alias SkateWeb.AuthManager
 
   plug(:laboratory_features)
@@ -10,7 +10,7 @@ defmodule SkateWeb.PageController do
     username = AuthManager.Plug.current_resource(conn)
     _ = Logger.info("uid=#{username}")
 
-    settings = Settings.get_or_create(username)
+    settings = UserSettings.get_or_create(username)
 
     conn
     |> assign(:username, username)
